Plan 13 moduł etapy (#7)
* Added CRUD process

* Added Test for orders

+++ src/main/java/edu/pjatk/planista/order/dto/OrderResponse.java
@@ -12,7 +12,7 @@ public record OrderResponse(
         String product,
         LocalDate dateFrom,
         LocalDate dateTo,
-        Long quantity,
+        Integer quantity,
         Instant createdAt,
         Instant updatedAt,
         DictItemDto status,
